Changed baudrate to 921600, enabled pppd package, enabled dropbear and added simple /root/ppp script.
This allows me to have a network connection over the serial link. For this, just log into the minidisc system, execute ./ppp and close the terminal software.
Then on the PC you execute "sudo pppd debug nodetach noauth nocrtscts local 192.168.10.1:192.168.10.10 /dev/ttyUSB0 921600" to create the network connection.
After this, you should be able to do "ssh root@192.168.10.10" and connect to the device.